getFarmWithTradingFeesApy.ts
