+++ src/api/stats/avax/getLydLpApys.js
@@ -8,7 +8,7 @@ const { compound } = require('../../../utils/compound');
 const { AVAX_CHAIN_ID, BASE_HPY } = require('../../../constants');
 const getBlockNumber = require('../../../utils/getBlockNumber');
 const { getTradingFeeApr } = require('../../../utils/getTradingFeeApr');
-const getFarmWithTradingFeesApy = require('../../../utils/getFarmWithTradingFeesApy');
+import { getFarmWithTradingFeesApy } from '../../../utils/getFarmWithTradingFeesApy';
 const { lydiaClient } = require('../../../apollo/client');
 
 const ERC20 = require('../../../abis/ERC20.json');
