+++ src/api/stats/bsc/degens/getSwampApys.js
@@ -9,7 +9,7 @@ const { getTotalLpStakedInUsd } = require('../../../../utils/getTotalStakedInUsd
 const { BSC_CHAIN_ID, BASE_HPY } = require('../../../../constants');
 const getBlockNumber = require('../../../../utils/getBlockNumber');
 const { getTradingFeeApr } = require('../../../../utils/getTradingFeeApr');
-const getFarmWithTradingFeesApy = require('../../../../utils/getFarmWithTradingFeesApy');
+import { getFarmWithTradingFeesApy } from '../../../../utils/getFarmWithTradingFeesApy';
 const { cakeClient } = require('../../../../apollo/client');
 
 const masterchef = '0x33AdBf5f1ec364a4ea3a5CA8f310B597B8aFDee3';
