+++ src/api/stats/bsc/omnifarm/getOmnifarmApys.js
@@ -7,7 +7,7 @@ const { compound } = require('../../../../utils/compound');
 const { BSC_CHAIN_ID, BASE_HPY } = require('../../../../constants');
 const { getTotalLpStakedInUsd } = require('../../../../utils/getTotalStakedInUsd');
 const { getTradingFeeApr } = require('../../../../utils/getTradingFeeApr');
-const getFarmWithTradingFeesApy = require('../../../../utils/getFarmWithTradingFeesApy');
+import { getFarmWithTradingFeesApy } from '../../../../utils/getFarmWithTradingFeesApy';
 const { cakeClient } = require('../../../../apollo/client');
 const getBlockNumber = require('../../../../utils/getBlockNumber');
 const pools = require('../../../../data/degens/omnifarmLpPools.json');
